# Conflicts:
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/fragment/implementations/HomeFragment.java
#	v8engine/src/main/res/layout-land/wizard_page_one.xml
#	v8engine/src/main/res/layout-land/wizard_page_three.xml
#	v8engine/src/main/res/layout-land/wizard_page_two.xml
#	v8engine/src/main/res/layout-large/wizard_page_one.xml
#	v8engine/src/main/res/layout-large/wizard_page_three.xml
#	v8engine/src/main/res/layout-large/wizard_page_two.xml
#	v8engine/src/main/res/layout-xlarge-land/wizard_page_one.xml
#	v8engine/src/main/res/layout-xlarge-land/wizard_page_three.xml
#	v8engine/src/main/res/layout-xlarge-land/wizard_page_two.xml
#	v8engine/src/main/res/layout-xlarge/wizard_page_one.xml
#	v8engine/src/main/res/layout-xlarge/wizard_page_three.xml
#	v8engine/src/main/res/layout-xlarge/wizard_page_two.xml
#	v8engine/src/main/res/layout/wizard_page_one.xml
#	v8engine/src/main/res/layout/wizard_page_three.xml
#	v8engine/src/main/res/layout/wizard_page_two.xml

+++ model/src/main/java/cm/aptoide/pt/model/v3/InAppBillingSkuDetailsResponse.java
@@ -31,7 +31,7 @@ import lombok.EqualsAndHashCode;
 
     @JsonProperty("productId") private String productId;
     private String price;
-    @JsonProperty("price_amount") private float priceAmount;
+    @JsonProperty("price_amount") private double priceAmount;
     private String currency;
     private String title;
     private String description;
